move KeyRangUtils to TiKey
