+++ src/main/java/com/pingcap/tikv/operation/ScanIterator.java
@@ -27,14 +27,13 @@ import com.pingcap.tikv.meta.TiKey;
 import com.pingcap.tikv.region.RegionManager;
 import com.pingcap.tikv.region.RegionStoreClient;
 import com.pingcap.tikv.region.TiRegion;
-import com.pingcap.tikv.util.KeyRangeUtils;
 import com.pingcap.tikv.util.Pair;
 import java.util.Iterator;
 import java.util.List;
 import java.util.NoSuchElementException;
 
 public class ScanIterator implements Iterator<Kvrpcpb.KvPair> {
-  private final Range keyRange;
+  private final Range<TiKey> keyRange;
   private final int batchSize;
   protected final TiSession session;
   private final RegionManager regionCache;
@@ -54,7 +53,7 @@ public class ScanIterator implements Iterator<Kvrpcpb.KvPair> {
       long version) {
     this.startKey = startKey;
     this.batchSize = batchSize;
-    this.keyRange = KeyRangeUtils.toRange(range);
+    this.keyRange = TiKey.toRange(range);
     this.session = session;
     this.regionCache = rm;
     this.version = version;
@@ -110,7 +109,6 @@ public class ScanIterator implements Iterator<Kvrpcpb.KvPair> {
     return true;
   }
 
-  @SuppressWarnings("unchecked")
   private boolean contains(ByteString key) {
     return keyRange.contains(new TiKey<>(key));
   }
