--- src/main/java/com/pingcap/tikv/util/KeyRangeUtils.java
@@ -1,59 +1,0 @@
-/*
- * Copyright 2017 PingCAP, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.pingcap.tikv.util;
-
-import com.google.common.collect.Range;
-import com.google.protobuf.ByteString;
-import com.pingcap.tikv.kvproto.Coprocessor;
-import com.pingcap.tikv.meta.TiKey;
-
-public class KeyRangeUtils {
-  public static Range toRange(Coprocessor.KeyRange range) {
-    if (range == null || (range.getStart().isEmpty() && range.getEnd().isEmpty())) {
-      return Range.all();
-    }
-    if (range.getStart().isEmpty()) {
-      return Range.lessThan(new TiKey<>(range.getEnd()));
-    }
-    if (range.getEnd().isEmpty()) {
-      return Range.atLeast(new TiKey<>(range.getStart()));
-    }
-    return Range.closedOpen(new TiKey<>(range.getStart()), new TiKey<>(range.getEnd()));
-  }
-
-  public static Range makeRange(ByteString startKey, ByteString endKey) {
-    if (startKey.isEmpty() && endKey.isEmpty()) {
-      return Range.all();
-    }
-    if (startKey.isEmpty()) {
-      return Range.lessThan(new TiKey<>(endKey));
-    } else if (endKey.isEmpty()) {
-      return Range.atLeast(new TiKey<>(startKey));
-    }
-    return Range.closedOpen(new TiKey<>(startKey), new TiKey<>(endKey));
-  }
-
-  static String formatByteString(ByteString key) {
-    StringBuilder sb = new StringBuilder();
-    for (int i = 0; i < key.size(); i++) {
-      sb.append(key.byteAt(i) & 0xff);
-      if (i < key.size() - 1) {
-        sb.append(",");
-      }
-    }
-    return sb.toString();
-  }
-}
